chore: 폴더 구조 통일
