+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 import { fetchNaverData } from '@/app/lib/api';
 import { Suspense } from 'react';
 import Loading from './loading';
-import ProductList from './product/ui/ProductList';
+import ProductList from './product/ProductList';
 
 const fetchForSSG = async () => await fetchNaverData();
