+++ src/app/product/[id]/page.tsx
@@ -1,5 +1,5 @@
 import Loading from '@/app/loading';
-import { Detail } from '@/app/product/[id]/ui';
+import Detail from '@/app/product/[id]/components/Detail';
 import { Suspense } from 'react';
 
 export default async function DetailPage() {
