+++ src/app/product/[id]/ui/index.ts
@@ -1,3 +1,2 @@
-export { default as Detail } from './Detail';
 export { default as Option } from './Option';
 export { default as DetailView } from './DetailView';
