+++ src/app/product/ui/index.ts
@@ -1,2 +1,2 @@
-export { default as ProductList } from './ProductList';
+export { default as ProductList } from '../ProductList';
 export { default as ProductListView } from './ProductListView';
